% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{chatter.auth}
\alias{chatter.auth}
\title{An experimental and unofficial R package for interacting with
openai-gpt}
\usage{
chatter.auth(openai_secret_key = NULL)
}
\arguments{
\item{openai_secret_key}{The OpenAI secret key for authentication}
}
\value{
A response from openai-gpt.
}
\description{
This R package provides an interface for interacting with openai-gpt.
It allows users to send messages to openai-gpt and receive responses.
}
\details{
This package provides a simple interface for interacting with openai.
gptchatteR is an R package that provides a simple interface for
interacting with the openai GPT-3 language processing model.
With gptchatteR, users can easily send messages to GPT-3 and receive
human-like responses in return. gptchatteR is easy to use and requires
no prior knowledge of GPT-3 or natural language processing. Simply
install the package, authenticate with your openAI API key, and start
sending messages to GPT-3. The package also includes functions for
processing and displaying the generated responses.
}
\examples{
library(gptchatteR)

# Create a chatter using your OpenAI API key
chatter.auth(openai_secret_key = "sk-qGTnjsCI8mZkCtvXVe6S")

chatter.create()

# Start talking!
chatter.chat("Hi!")

# Feed the chatter with prior information
chatter.feed("My name is Isin.")
chatter.chat("What is my name?")


# A real example (Note: The object (i.e. df) should be defined.)
chatter.feed("I have a dataframe named df. It has two columns: A and B")

cplot <- chatter.plot("use ggplot to plot a scatterplot where x axis is
A and y axis is B", run = FALSE)


# Print the plot (if you are lucky and it runs properly)
cplot$plot




Authenticate chatter with an OpenAI secret key

This function can be used to authenticate a user with an OpenAI
secret key. It takes the argument \code{openai_secret_key} and if it
i snot provided, an error message is displayed with instructions on
how to get an OpenAI secret key. The secret key is then stored as an
environment variable.

chatter.auth(openai_secret_key = "lbkFJuMOwtvGYKOWzYJsFNw2L")

}
\author{
Isin Altinkaya
}
