% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{chatter.chat}
\alias{chatter.chat}
\title{This function allows the user to chat with a given model.}
\usage{
chatter.chat(input, echo = FALSE, return_response = FALSE, feed = FALSE, ...)
}
\arguments{
\item{input}{A character string containing the text to be sent to the model.}

\item{echo}{A logical indicating whether to echo the input in the response.}

\item{return_response}{A logical indicating whether to return the response object.}

\item{feed}{A logical indicating whether to feed the input to the model.}

\item{...}{Further arguments passed to openai::create_completion.}
}
\value{
If \code{return_response = TRUE}, a list containing the response from openai::create_completion.
Otherwise nothing is returned.
}
\description{
This function allows the user to chat with a given model.
}
\examples{
chatter.chat("Hi there!")

}
\references{
OpenAI, GPT-2 (2019). \url{https://openai.com/blog/better-language-models/}
}
\seealso{
\code{\link[openai]{create_completion}}
}
