% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{chatter.create}
\alias{chatter.create}
\title{Create an OpenAI chatterbot}
\usage{
chatter.create(
  model = "text-davinci-003",
  temperature = 0.5,
  max_tokens = 100,
  ...
)
}
\arguments{
\item{model}{The OpenAI model to use (default is \code{text-davinci-003})}

\item{temperature}{Float between 0 and 1, representing the degree of
randomness (default is \code{0.5})}

\item{max_tokens}{Maximum number of tokens to generate
(default is \code{100})}

\item{...}{Other arguments passed to \code{\link[openai]{create_completion}}}
}
\value{
A \code{chatter} object containing the OpenAI secret key and the
provided parameters
}
\description{
Create an OpenAI chatterbot
}
\examples{

chatter.auth(openai_secret_key = "MY_SECRET_KEY")

}
\seealso{
\code{\link{chatter.auth}, \link{chatter.feed}},
\code{\link{chatter.chat}},\code{\link{chatter.plot}}
}
