% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{chatter.feed}
\alias{chatter.feed}
\title{This function enables users to chat with the chatterbot.}
\usage{
chatter.feed(new_input)
}
\arguments{
\item{input}{A string of text inputted by the user.}

\item{echo}{A logical indicating whether to echo the input.}

\item{return_response}{A logical indicating whether to return the response item (default is to return the response as text).}

\item{feed}{A logical indicating whether to feed the input to the chatterbot and chat at the same time.}

\item{...}{Other parameters to pass to openai::create_completion.}
}
\value{
Returns the response if return_response = TRUE otherwise nothing is returned.
}
\description{
This function enables users to chat with the chatterbot.
}
\examples{
chatter.chat("Hello, how are you?")
}
