% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{chatter.plot}
\alias{chatter.plot}
\title{This function enables users to plot graphs with the chatterbot.}
\usage{
chatter.plot(input, echo = FALSE, feed = FALSE, run = FALSE, ...)
}
\arguments{
\item{input}{A string of text inputted by the user.}

\item{echo}{A logical indicating whether to echo the input.}

\item{feed}{A logical indicating whether to feed the input to the chatterbot and plot at the same time.}

\item{run}{A logical indicating whether to run the plot command or return it as a chatterplot object.}

\item{...}{Other parameters to pass to openai::create_completion.}
}
\value{
Returns the plot command as a chatterplot object if run = FALSE, otherwise nothing is returned.
}
\description{
This function enables users to plot graphs with the chatterbot.
}
\examples{
chatter.plot("Draw a line graph of x and y.")
}
